feat: add custom css

+++ app/FileSystem.php
@@ -7,8 +7,26 @@ class FileSystem
 
     protected const BLACKLISTED_FOLDER = '/\/vendor$/';
 
+    public static function createFolder(string $dirPath)
+    {
+        $absolutePath = getcwd() . '/' . $dirPath;
+        if (!is_dir($absolutePath)) {
+            exec("mkdir -p \"$absolutePath\"");
+        }
+    }
+
+    public static function emptyFolder(string $dirPath)
+    {
+        self::delete($dirPath, false);
+    }
+
+    public static function deleteFolder(string $dirPath)
+    {
+        self::delete($dirPath, true);
+    }
+
     // see: https://www.tutorialspoint.com/how-to-recursively-delete-a-directory-and-its-entire-contents-files-plus-sub-dirs-in-php
-    public static function delete($dirPath)
+    private static function delete(string $dirPath, bool $selfDelete)
     {
         if (is_dir($dirPath)) {
             $files = scandir($dirPath);
@@ -16,13 +34,15 @@ class FileSystem
                 if ($file !== '.' && $file !== '..') {
                     $filePath = $dirPath . '/' . $file;
                     if (is_dir($filePath)) {
-                        self::delete($filePath);
+                        self::delete($filePath, true);
                     } else {
                         unlink($filePath);
                     }
                 }
             }
-            rmdir($dirPath);
+            if ($selfDelete) {
+                rmdir($dirPath);
+            }
         }
     }
 
@@ -54,5 +74,17 @@ class FileSystem
         }
 
         return $res;
+    }
+
+    public static function readFile($filePath)
+    {
+        $absolutePath = getcwd() . '/' . $filePath;
+        return file_get_contents($absolutePath);
+    }
+
+    public static function writeFile($filePath, $fileContent)
+    {
+        $absolutePath = getcwd() . '/' . $filePath;
+        file_put_contents($absolutePath, $fileContent);
     }
 }
